Renamed cryptoController
